Re-implement the audit logging feature.
This now tracks the ID of the object that has been changed, and the full state of the object before and after.

Unfortunately there was no way I could retain backwards compatibility with the old model so I've had to truncate the `audit_records` table as part of one of the migrations.

This will mean a MAJOR version bump I guess - might as well make audit records default to on as part of that release.

resolves #22 and #27